livekit-deepgram-service/requirements.txt
livekit
livekit-agents
livekit-plugins-deepgram
python-dotenv
httpx
pyaudio
sounddevice
soundfile
